requirements.txt
requests
pymongo
pandas
kafka-python==2.0.2
numpy
python-dotenv